requirements.txt
langgraph
langchain
langchain-google-genai
streamlit
mcp
python-dotenv
pandas
pydantic
